(medium)[Add]: World discovery metrics now logged
Why: Pycolab environment tracks visitation metrics, but don't actually get locked in tensorboard. This is a separate branch because tracking these in TrajInfo is not generalizable across environments, and should be a temporary addition for these experiments.

---------------------------------------------------------------------------------------------------------

How: TrajInfo stores visitation frequencies and runs the logic for determining first visit time for each object in the 5 room maze. Note that changing the maze requires changing the logic in TrajInfo (you'll have to add variables to track additional objects manually).

---------------------------------------------------------------------------------------------------------

Tags: #pycolab #metrics

References: