(minor)[Fix,Refactor]: Forward model losses added up in compute_loss()
Why: Before, a list of forward model losses was being passed out of the disagreement model. This was so
these losses could be logged individually before being summed. This method was not working in tests, and
some weird learning was occurring (including the inverse model, which should be unchanged).

---------------------------------------------------------------------------------------------------------

How: Forward model losses are computed and immediately summed in the curiosity algorithm compute_loss()
method. This allows for further clarity and easier debugging. Forward losses are not individually logged,
although their sum is still logged. Also, we should check but we seemed to fix a bug where the first
forward model loss was added twice to the list of forward losses, which would explain why this loss was
always much larger than the other four.

---------------------------------------------------------------------------------------------------------

Tags:

References: