(minor)[Reformat]: Small changes to way disagreement is computed
Why: The only difference between ICM and Disagreement is multiple forward models. The bug has to be
occurring in how we generate the reward in Disagreement.

---------------------------------------------------------------------------------------------------------

How: Took out keep_dims=True when calculating variance - it seems this does not matter if you specify
the axis along which to take the variance. Additionally, separate out a list of forward predictions and
a stacked version of this list. The stacked version does not get backpropped through, and is just used
to calculate variance and generate reward. The list of individual features should get backpropped through,
with each feature getting its own loss backpropagated through itself.

---------------------------------------------------------------------------------------------------------

Tags:

References: