(major)[Add]: Added NDIGO-1 algorithm
Why: Need to add the NDIGO-1 algorithm to test, before being able to expand to NDIGO-k.

---------------------------------------------------------------------------------------------------------

How: NDIGO-1 is implemented by making very minor modifications (if clauses) to the sampler. This is the
first step towards a multi-curiosity alg codebase (with algs other than ICM), so a few paradigms are
started here (not final). A separate NDIGO recurrent agent extends the usual RecurrentPgAgent in order
to keep track of previous losses for each environment, and the previous gru cell state. Btw, the previous
gru cell state is updated every time the agent runs curiosity_step (this means env 1 could update it and
then this gru cell state is used by env 3 for its next step which is kind of weird). This is how the
functional LSTM policy is updated (the lstm state is updated at every agent.step call, even across envs).
This code will be used to test NDIGO-1 with a single forward predictor (as opposed to 10), and gru cell
state being updated at every step across environments (as opposed to keeping track of the last forward
state for every environment instance).

---------------------------------------------------------------------------------------------------------

Tags: #ndigo

References:
1) NDIGO Paper: [https://arxiv.org/pdf/1902.07685.pdf]