(major)[Add]: Changed NDIGO to batch processing
Why: Batch processing eliminates the need for complicated buffers/machinery that you'd need to do
multistep rewarding at every step.

---------------------------------------------------------------------------------------------------------

How: Sampling is virtually unchanged, except that intrinsic rewards are zeroed out for NDIGO. Then, an
entire batch is processed and multistep rewards are generated as well as losses. The losses generated
are the same as they would be if we were running sequential intrinsic reward generation. This can run on
Pycolab only at the moment while we test.

---------------------------------------------------------------------------------------------------------

Tags: #ndigo

References: