(minor)[Fix]: Fixed pycolab metric logging
Why: For some reason (partially because pycolab env was not being mounted) the pycolab metrics were not
being correctly logged. Also, had to clean up some of the sequential NDIGO remnants.

---------------------------------------------------------------------------------------------------------

How: Slight changes to PycolabTrajInfo, and took out the CpuNdigoWaitResetCollector and its import in
the launch file.

---------------------------------------------------------------------------------------------------------

Tags:

References: