(minor)[Fix]: Took out last ReLU in forward model
Why: Output of the forward model is a predicted image (probability distribution). This is then passed
through a sigmoid in the BCELoss function. This sigmoid takes (-inf, inf) inputs and maps them to (0,1).
Mapping the last layer through ReLU means your probabilities are always above 0.5, which is incorrect.

---------------------------------------------------------------------------------------------------------

How: The last ReLU is taken out of the forward model. The forward model now resembles ICM's inverse
model - just two linear layers with a ReLU in between. BCELoss with logits is used, as it's more
numerically stable than BCELoss with a sigmoid function applied separately.

---------------------------------------------------------------------------------------------------------

Tags:

References: