(major)[Add,Fix]: NDIGO Forward models taken out of lists, grad visualizations added
Why: Gradient and computation graph visualizations were required for debugging. In order for the forward
model to train appropriately you need to take the forward models out of lists.

---------------------------------------------------------------------------------------------------------

How: Each forward model is instantiated and updated one by one. Code is lengthy, but we can maybe fix
this later if it works. Gradient flow visualizer plotting was added in utils, as well as computation
graph visualizations in utils too.

---------------------------------------------------------------------------------------------------------

Tags: #visualization

References: