(major)[Add]: Integrated custom fetch environment
Why: Planar fetch push environment is now integrated, and runs with the PPO/ICM pipeline. Images are being
rendered correctly.

---------------------------------------------------------------------------------------------------------

How: Custom environment files for fetch are kept in rlpyt/envs. New cameras are defined, and an action
wrapper is used to constrain actions to move in a 2D plane along the table. The pipeline works with ICM
and PPO, but the images are too big to store large batches in memory. Need to figure out a way to down
sample as we go forward. Also, rewards are modified to be between 0 and 1 and not -1 and 0 when sparse.
The GymWrapper defined by rlpyt has lines commented out such that the process does not hang when starting
environments - for some reason calling reset and step in the init function here causes issues. A custom
encoder that works with 500x500 images is implemented. Minor changes to the observation space for the
original Fetch environment have been made, such that it is no longer a GoalEnv but a standard Env class
that can return either a spaces.Dict observation with joint and velocity state or a standard RGB image
observation as a spaces.Box.

---------------------------------------------------------------------------------------------------------

Tags:

References: