(minor)[Fix]: Switched back to 4 models for Disagreement
Why: 5 models was running into swap memory limits on certain machines when running 128 environments
and RGB Pycolab. 4 models has been shown to work just as well as 5 on Atari, but we should test this
on Pycolab as well in the future.

---------------------------------------------------------------------------------------------------------

How: Took out the 5th forward model in Disagreement.

---------------------------------------------------------------------------------------------------------

Tags:

References: